refactor: improve proposal dashboard
