+++ src/app/datasets/dashboard/dashboard.component.html
@@ -1,8 +1,10 @@
 <breadcrumb></breadcrumb>
 <full-text-search-bar
   #searchBar
-  data-cy="text-search"
+  data-cy="dataset-text-search"
   [placeholder]="'Text Search'"
+  (textChange)="onTextChange($event)"
+  (searchAction)="onSearchAction()"
 >
 </full-text-search-bar>
 <mat-sidenav-container [hasBackdrop]="false">
