+++ src/app/proposals/proposals.module.ts
@@ -29,7 +29,7 @@ import { FlexLayoutModule } from "@ngbracket/ngx-layout";
 import { FileSizePipe } from "shared/pipes/filesize.pipe";
 import { ProposalFilterComponent } from "./proposal-filter/proposal-filter.component";
 import { LogbooksModule } from "logbooks/logbooks.module";
-import { ProposalDashboardComponent } from "./proposal-dashboard/proposal-dashboard.component";
+import { ProposalTableComponent } from "./proposal-table/proposal-table.component";
 import { MatDatepickerModule } from "@angular/material/datepicker";
 import { LogbookEffects } from "state-management/effects/logbooks.effects";
 import { logbooksReducer } from "state-management/reducers/logbooks.reducer";
@@ -37,6 +37,7 @@ import { ProposalLogbookComponent } from "./proposal-logbook/proposal-logbook.co
 import { RelatedProposalsComponent } from "./related-proposals/related-proposals.component";
 import { ProposalDatasetsComponent } from "./proposal-datasets/proposal-datasets.component";
 import { TranslateModule } from "@ngx-translate/core";
+import { ProposalDashboardComponent } from "./proposal-dashboard/proposal-dashboard.component";
 
 @NgModule({
   imports: [
@@ -64,10 +65,11 @@ import { TranslateModule } from "@ngx-translate/core";
     TranslateModule,
   ],
   declarations: [
+    ProposalDashboardComponent,
     ViewProposalPageComponent,
     ProposalDetailComponent,
     ProposalFilterComponent,
-    ProposalDashboardComponent,
+    ProposalTableComponent,
     ProposalLogbookComponent,
     RelatedProposalsComponent,
     ProposalDatasetsComponent,
