+++ src/app/shared/modules/full-text-search-bar/full-text-search-bar.component.scss
