+++ src/app/shared/modules/full-text-search-bar/full-text-search-bar.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from "@angular/core";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { MatButtonModule } from "@angular/material/button";
+import { MatCardModule } from "@angular/material/card";
+import { MatFormFieldModule } from "@angular/material/form-field";
+import { MatIconModule } from "@angular/material/icon";
+import { MatInputModule } from "@angular/material/input";
+import { MatSelectModule } from "@angular/material/select";
+import { FullTextSearchBarComponent } from "./full-text-search-bar.component";
+
+@NgModule({
+  declarations: [FullTextSearchBarComponent],
+  imports: [
+    MatFormFieldModule,
+    MatSelectModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatInputModule,
+    MatIconModule,
+    MatCardModule,
+    MatButtonModule,
+  ],
+  exports: [FullTextSearchBarComponent],
+})
+export class FullTextSearchBarModule {}
